+++ sharepasswordAzure/Program.cs
@@ -317,6 +317,17 @@ if (builder.Configuration.GetValue("Application:EnableHttpsRedirection", false))
 {
     app.UseHttpsRedirection();
 }
+
+app.Use(async (context, next) =>
+{
+    context.Response.Headers["Content-Security-Policy"] = "default-src 'self'; object-src 'none'; base-uri 'self'; form-action 'self'; frame-ancestors 'none'";
+    context.Response.Headers["X-Frame-Options"] = "DENY";
+    context.Response.Headers["X-Content-Type-Options"] = "nosniff";
+    context.Response.Headers["Referrer-Policy"] = "strict-origin-when-cross-origin";
+
+    await next();
+});
+
 app.UseRouting();
 
 app.UseAuthentication();